agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-strands>=0.1.1",
    "fastapi>=0.134.0",
    "strands-agents[mistral]>=1.28.0",
    "uvicorn>=0.41.0",
]
